requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
opencv-python-headless==4.10.0.84
paddlepaddle==2.6.2
paddleocr==2.9.1
onnxruntime==1.19.0
rapidfuzz==3.10.1
numpy==1.26.4
supabase>=2.10.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
